copy over the contents of the portfolio.md
i had initially written a few self motivation lines when i freshly created this file. but later on according to the tasks assigned to us on Archisacademy, i thought it wise to completely remove the self motivation lines and have the same contents in the portfolio file reside here as well.